+++ core/inference/functions.py
@@ -43,7 +43,7 @@ config_name = "model_index.json"
 
 
 torch_older_than_200 = version.parse(torch.__version__) < version.parse("2.0.0")
-torch_newer_or_same_as_210 = version.parse(torch.__version__) >= version.parse("2.1.0")
+torch_newer_or_same_as_210 = version.parse(torch.__version__) > version.parse("2.0.1")
 
 
 def is_ipex_available():
