+++ core/inference/pytorch/latents.py
